refactor: move pena profile slice into core application
