+++ backend/src/api/dependencies/use_cases.py
@@ -8,6 +8,9 @@ from core.application.use_cases.get_player_profile_usecase import GetPlayerProfi
 from core.application.use_cases.get_season_match_insights_usecase import (
     GetSeasonMatchInsightsUseCase,
 )
+from core.application.use_cases.manage_pena_profile_usecase import (
+    ManagePenaProfileUseCase,
+)
 from core.application.use_cases.update_player_profile_usecase import (
     UpdatePlayerProfileUseCase,
 )
@@ -25,9 +28,6 @@ from persistence.application.use_cases.manage_pena_labels_usecase import (
 )
 from persistence.application.use_cases.manage_pena_membership_usecase import (
     ManagePenaMembershipUseCase,
-)
-from persistence.application.use_cases.manage_pena_profile_usecase import (
-    ManagePenaProfileUseCase,
 )
 from persistence.application.use_cases.manage_pena_seasons_usecase import (
     ManagePenaSeasonsUseCase,
