+++ backend/src/core/application/ports/__init__.py
@@ -1,10 +1,12 @@
 from .nationality_query_port import NationalityQueryPort
+from .pena_profile_port import PenaProfilePort
 from .pena_query_port import PenaQueryPort
 from .player_profile_port import PlayerProfilePort
 from .season_match_insights_port import SeasonMatchInsightsPort
 
 __all__ = [
     "NationalityQueryPort",
+    "PenaProfilePort",
     "PenaQueryPort",
     "PlayerProfilePort",
     "SeasonMatchInsightsPort",
