+++ backend/src/core/application/use_cases/__init__.py
@@ -2,6 +2,12 @@ from .get_nationalities_usecase import GetNationalitiesUseCase
 from .get_penas_usecase import GetPenasUseCase
 from .get_player_profile_usecase import GetPlayerProfileUseCase
 from .get_season_match_insights_usecase import GetSeasonMatchInsightsUseCase
+from .manage_pena_profile_usecase import (
+    InvalidPenaProfileImageError,
+    ManagePenaProfileUseCase,
+    PenaProfileAccessDeniedError,
+    PenaProfileNotFoundError,
+)
 from .season_match_insights_errors import (
     InvalidSeasonInsightsDataError,
     PenaSeasonNotFoundError,
@@ -20,10 +26,14 @@ __all__ = [
     "GetPenasUseCase",
     "GetPlayerProfileUseCase",
     "GetSeasonMatchInsightsUseCase",
+    "InvalidPenaProfileImageError",
     "InvalidNationalityError",
     "InvalidPlayerUpdateDataError",
     "InvalidProfileImageError",
     "InvalidSeasonInsightsDataError",
+    "ManagePenaProfileUseCase",
+    "PenaProfileAccessDeniedError",
+    "PenaProfileNotFoundError",
     "PenaSeasonNotFoundError",
     "PenaSeasonPenaNotFoundError",
     "PlayerUpdate",
