+++ backend/src/persistence/application/use_cases/__init__.py
@@ -1,10 +1,16 @@
 from core.application.models import PenaInfo as PenaSummary
-from core.application.models import PenasPage, PlayerProfile
+from core.application.models import PenaProfileInfo, PenaProfileUpdate, PenasPage, PlayerProfile
 from core.application.use_cases.get_nationalities_usecase import GetNationalitiesUseCase
 from core.application.use_cases.get_penas_usecase import GetPenasUseCase
 from core.application.use_cases.get_player_profile_usecase import GetPlayerProfileUseCase
 from core.application.use_cases.get_season_match_insights_usecase import (
     GetSeasonMatchInsightsUseCase,
+)
+from core.application.use_cases.manage_pena_profile_usecase import (
+    InvalidPenaProfileImageError,
+    ManagePenaProfileUseCase,
+    PenaProfileAccessDeniedError,
+    PenaProfileNotFoundError,
 )
 from core.application.use_cases.update_player_profile_usecase import (
     InvalidNationalityError as PlayerInvalidNationalityError,
@@ -71,14 +77,6 @@ from .manage_pena_membership_usecase import (
     PenaMembershipPlayerNotFoundError,
     PenaMembershipUpdate,
     PenaMembershipUserProfileNotFoundError,
-)
-from .manage_pena_profile_usecase import (
-    InvalidPenaProfileImageError,
-    ManagePenaProfileUseCase,
-    PenaProfileAccessDeniedError,
-    PenaProfileInfo,
-    PenaProfileNotFoundError,
-    PenaProfileUpdate,
 )
 from .manage_season_competition_usecase import (
     ManageSeasonCompetitionUseCase,
